feat(FooterContainer): rename comp
